add code for id

+++ ms/src/main/java/org/game/ms/player/Player.java
@@ -18,6 +18,6 @@ import org.game.ms.role.Role;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Player extends Role {
 
-    private long experience;
+    private int experience;
     private List<Profession> profession = new ArrayList<>();
 }
